+++ R/hgwr.R
@@ -151,7 +151,7 @@ hgwr.data.frame <- function(
 #' @describeIn hgwr Fit a HGWR model
 #' @export 
 hgwr_fit <- function(
-    formula, data, coords, bw = "CV",
+    formula, data, coords, bw = c("CV", "AIC"),
     kernel = c("gaussian", "bisquared"),
     alpha = 0.01, eps_iter = 1e-6, eps_gradient = 1e-6,
     max_iters = 1e6, max_retries = 1e6,
@@ -184,21 +184,25 @@ hgwr_fit <- function(
     if (model_desc$intercept$local) g <- cbind(1, g)
 
     ### Get bandwidth value
-    if (is.character(bw) && bw == "CV") {
+    if (is.character(bw)) {
+        bw <- match.arg(bw)
         bw_value <- NA_real_
         optim_bw <- TRUE
+        bw_criterion <- switch(bw, "CV" = 0L, "AIC" = 1L)
     } else if (is.numeric(bw) || is.integer(bw)) {
         bw_value <- bw
         optim_bw <- FALSE
+        bw_criterion <- -1L
     } else {
         bw_value <- Inf
-        optim_bw <- FALSE
+        optim_bw <- TRUE
+        bw_criterion <- 0L
     }
 
     ### Call C
     hgwr_result <- tryCatch({
         hgwr_bfml(
-            g, x, z, y, as.matrix(coords), group_index, bw_value, kernel_index,
+            g, x, z, y, as.matrix(coords), group_index, bw_value, bw_criterion, kernel_index,
             alpha, eps_iter, eps_gradient,
             as.integer(max_iters), as.integer(max_retries),
             as.integer(ml_type), as.integer(verbose)
@@ -210,17 +214,8 @@ hgwr_fit <- function(
         bw_value <- hgwr_result$bw
 
     ### Prepare Return Result
-    result <- list(
-        gamma = hgwr_result$gamma,
-        beta = hgwr_result$beta,
-        mu = hgwr_result$mu,
-        D = hgwr_result$D,
-        sigma = hgwr_result$sigma,
-        bw = bw_value,
-        logLik = hgwr_result$logLik,
-        trS = hgwr_result$trS,
-        var_beta = hgwr_result$var_beta,
-        effects = list(
+    result <- c(hgwr_result, list(
+            effects = list(
             global.fixed = gfe,
             local.fixed = lfe,
             random = model_desc$random.effects,
@@ -238,7 +233,8 @@ hgwr_fit <- function(
         ),
         groups = group_unique,
         coords = coords
-    )
+    ))
+    result$bw <- bw_value
     class(result) <- "hgwrm"
     result
 }
@@ -329,6 +325,8 @@ residuals.hgwrm <- function(object, ...) {
 #' If it is set to `FALSE`, the test will not be executed.
 #' If it is set to `TRUE`, the test will be executed with default parameters (see details below).
 #' It accepts a list to enable the test with specified parameters.
+#' @param verbose An Integer value to control whether additional messages
+#' during testing spatial heterogeneity should be reported.
 #'
 #' @return A list containing summary informations of this `hgwrm` object
 #' with the following fields.
@@ -344,12 +342,25 @@ residuals.hgwrm <- function(object, ...) {
 #'  \item{\code{bw}}{Bandwidth (unit: number of nearest neighbours) used to make spatial kernel density estimation. Default: `10`.}
 #'  \item{\code{poly}}{The number of polynomial terms used in the local polynomial estimation. Default: `2`.}
 #'  \item{\code{resample}}{Total resampling times. Default: `5000`.}
+#'  \item{\code{kernel}}{The kernel function used in the local polynomial estimation. Options are `"gaussian"` and `"bisquared"`. Default: `"bisquared"`.}
 #' }
+#' 
+#' @examples 
+#' data(multisampling)
+#' m <- hgwr(
+#'  formula = y ~ L(g1 + g2) + x1 + (z1 | group),
+#'  data = multisampling$data,
+#'  coords = multisampling$coords,
+#'  bw = 10
+#' )
+#' summary(m)
+#' summary(m, test_hetero = TRUE)
+#' summary(m, test_hetero = list(kernel = "gaussian"))
 #' 
 #' @importFrom stats AIC logLik pt sd
 #' @seealso [hgwr()].
 #' @export 
-summary.hgwrm <- function(object, ..., test_hetero = FALSE) {
+summary.hgwrm <- function(object, ..., test_hetero = FALSE, verbose = 0) {
     if (!inherits(object, "hgwrm")) {
         stop("It's not a hgwrm object.")
     }
@@ -376,8 +387,7 @@ summary.hgwrm <- function(object, ..., test_hetero = FALSE) {
     ### Significance test
     significance <- list()
     #### Beta
-    enp <- attr(logLik_object, "df")
-    edf <- length(y) - enp
+    edf <- object$edf
     se_beta <- sqrt(object$var_beta)
     t_beta <- abs(object$beta) / se_beta
     p_beta <- (1 - stats::pt(t_beta, edf)) * 2
@@ -387,25 +397,52 @@ summary.hgwrm <- function(object, ..., test_hetero = FALSE) {
         tv = t_beta,
         pv = p_beta
     )
-    #### Gamma
-    if (test_hetero == TRUE || is.list(test_hetero)) {
+    #### Gamma different from 0
+    t_gamma <- object$gamma / object$gamma_se
+    p_gamma <- stats::pt(-abs(object$gamma / object$gamma_se), object$edf) * 2
+    significance$gamma <- list(
+        tv = t_gamma,
+        pv = p_gamma
+    )
+    #### Gamma spatial heterogeneity
+    is_test_hetero <- FALSE
+    if (length(test_hetero) == 1 && is.logical(test_hetero) && test_hetero == TRUE) {
+        is_test_hetero <- TRUE
+    } else if (is.list(test_hetero)) {
+        is_test_hetero <- TRUE
+    }
+    if (is_test_hetero) {
         bw <- 10L
         resample <- 5000L
         poly <- 2L
+        kernel <- "bisquared"
         if (is.list(test_hetero)) {
             bw <- ifelse("bw" %in% names(test_hetero), test_hetero$bw, bw)
             resample <- ifelse("resample" %in% names(test_hetero), test_hetero$resample, resample)
             poly <- ifelse("poly" %in% names(test_hetero), test_hetero$poly, poly)
+            kernel <- ifelse("kernel" %in% names(test_hetero), test_hetero$kernel, kernel)
         }
-        mean_gamma <- colMeans(object$gamma)
+        kernel <- match.arg(kernel, c("gaussian", "bisquared"))
+        kernel_id <- switch(kernel, "gaussian" = 0, "bisquared" = 1)
         sd_gamma <- apply(object$gamma, 2, stats::sd)
-        t_gamma <- spatial_hetero_perm(object$gamma, as.matrix(object$coords), poly = poly, resample = resample, bw = bw)
+        t_gamma <- spatial_hetero_perm(
+            object$gamma,
+            as.matrix(object$coords),
+            poly = poly,
+            resample = resample,
+            bw = bw,
+            kernel = kernel_id,
+            verbose = as.integer(verbose)
+        )
+        t_gamma$t <- sqrt(t_gamma$t)
+        t_gamma$t0 <- sqrt(t_gamma$t0)
         pv <- sapply(seq_along(t_gamma$t0), function(i) {
             with(t_gamma, mean(t[,i] > t0[i]))
         })
-        significance$gamma <- data.frame(
-            mean = mean_gamma,
+        significance$gamma$hetero <- list(
             sd = sd_gamma,
+            t = t_gamma$t,
+            t0 = t_gamma$t0,
             pv = pv
         )
     }
@@ -476,7 +513,7 @@ print.hgwrm <- function(x, decimal.fmt = "%.6f", ...) {
     )
     print.table.md(beta_str, ...)
     cat("\n")
-    cat("Local Fixed Effects", fill = T)
+    cat("Group-level Spatially Weighted Effects", fill = T)
     cat("-------------------", fill = T)
     cat("Bandwidth:", x$bw, "(nearest neighbours)", fill = T)
     gamma_fivenum <- t(apply(x$gamma, 2, fivenum))
@@ -486,11 +523,11 @@ print.hgwrm <- function(x, decimal.fmt = "%.6f", ...) {
     )
     print.table.md(gamma_str, ...)
     cat("\n")
-    cat("Random Effects", fill = T)
+    cat("Sample-level Random Effects", fill = T)
     cat("--------------", fill = T)
-    x_summary <- summary.hgwrm(x)
-    random_stddev <- x_summary$random.stddev
-    random_corr <- x_summary$random.corr
+    x <- summary.hgwrm(x)
+    random_stddev <- x$random.stddev
+    random_corr <- x$random.corr
     random_corr_str <- matrix2char(random_corr)
     random_corr_str[!lower.tri(random_corr)] <- ""
     random_corr_str <- rbind("", random_corr_str)
@@ -551,7 +588,7 @@ print.summary.hgwrm <- function(x, decimal.fmt = "%.6f", ...) {
     cat("\n")
 
     ### Parameter Estimates
-    cat("Parameter estimates", fill = T)
+    cat("Parameter Estimates", fill = T)
     cat("-------------------", fill = T)
     cat("Fixed effects:", fill = T)
     gfe <- x$effects$global.fixed
@@ -565,28 +602,84 @@ print.summary.hgwrm <- function(x, decimal.fmt = "%.6f", ...) {
         as.matrix(cbind(variable = gfe, x$significance$beta, stars = stars))
     ), ...)
     cat("\n")
-    cat("Local fixed effects:", fill = T)
+    cat("GLSW effects:", fill = T)
     lfe <- x$effects$local.fixed
     if (x$intercept$local) {
         lfe <- c("Intercept", lfe)
     }
-    gamma_stats <- matrix2char(t(apply(x$gamma, 2, fivenum)))
-    gamma_stats_name <- c("Min", "1st Quartile", "Median", "3rd Quartile", "Max")
-    if (!is.null(x$significance$gamma)) {
-        pv_lfe <- x$significance$gamma$pv
-        gamma_stats <- cbind(
-            gamma_stats,
-            pv = matrix2char(pv_lfe),
-            stars = vapply(pv_lfe, pv2stars, rep(" ", n = length(pv_lfe)))
-        )
-        gamma_stats_name <- c(gamma_stats_name, "Pr(>|t|)", "")
-    }
+    s_gamma <- t(apply(x$significance$gamma$pv, 2, function(p) { c(
+        "***" = mean(p < 0.001),
+        "**" = mean(p >= 0.001 & p < 0.01),
+        "*" = mean(p >= 0.01 & p < 0.05),
+        "." = mean(p >= 0.05 & p < 0.1)
+    ) })) * 100
+    rownames(s_gamma) <- lfe
+    gamma_stats <- cbind(
+        gamma_mean <- matrix2char(as.vector(colMeans(x$gamma))),
+        gamma_se_mean <- matrix2char(as.vector(colMeans(x$gamma_se))),
+        matrix2char(s_gamma, fmt = "%.1f%%")
+    )
+    gamma_stats_name <- c("Mean Est.", "Mean Sd.", colnames(s_gamma))
     gamma_str <- rbind(
         c("", gamma_stats_name),
         cbind(lfe, gamma_stats)
     )
     print.table.md(gamma_str, ...)
     cat("\n")
+    if (!is.null(x$significance$gamma$hetero)) {
+        cat("GLSW effect spatial heterogeneity:", fill = T)
+        h_gamma <- x$significance$gamma$hetero
+        pv_hetero <- h_gamma$pv
+        hetero_stats <- matrix2char(cbind(
+            gamma_sd = as.vector(h_gamma$sd),
+            t0 = as.vector(h_gamma$t0),
+            t.min = apply(h_gamma$t, 2, min),
+            t.max = apply(h_gamma$t, 2, max),
+            pv = pv_hetero
+        ))
+        hetero_stats <- cbind(
+            hetero_stats,
+            stars = vapply(pv_hetero, pv2stars, rep(" ", n = length(pv_hetero)))
+        )
+        hetero_stats_name <- c("Sd. Est.", "t0", "Min. t", "Max. t", "Pr(t>t0)", "")
+        hetero_str <- rbind(
+            c("", hetero_stats_name),
+            cbind(lfe, hetero_stats)
+        )
+        print.table.md(hetero_str, ...)
+        cat("\n")
+    }
+    cat("Bandwidth:", x$bw, "(nearest neighbours)", fill = T)
+    cat("\n")
+    cat("SLR effects:", fill = T)
+    random_stddev <- x$random.stddev
+    random_corr <- x$random.corr
+    random_corr_str <- matrix2char(random_corr)
+    random_corr_str[!lower.tri(random_corr)] <- ""
+    random_corr_str <- rbind("", random_corr_str)
+    random_corr_str[1, 1] <- "Corr"
+    re <- x$effects$random
+    if (x$intercept$random) {
+        re <- c("Intercept", re)
+    }
+    random_dev_str <- cbind(
+        "", re, matrix2char(cbind(colMeans(x$mu), matrix(random_stddev, ncol = 1)))
+    )
+    random_dev_str[1, 1] <- x$effects$group
+    random_dev_str <- rbind(
+        c("Groups", "Name", "Mean", "Std.Dev."),
+        random_dev_str
+    )
+    random_residual_str <- cbind(
+        matrix(c("Residual", "", sprintf(decimal.fmt, c(mean(x$residuals), x$sigma))), nrow = 1),
+        matrix("", nrow = 1, ncol = ncol(random_corr))
+    )
+    random_str <- rbind(
+        cbind(random_dev_str, random_corr_str),
+        random_residual_str
+    )
+    print.table.md(random_str, ...)
+    cat("\n", fill = T)
 
     ### Diagnostics
     cat("Diagnostics", fill = T)
@@ -599,7 +692,7 @@ print.summary.hgwrm <- function(x, decimal.fmt = "%.6f", ...) {
     cat("\n")
 
     ### Residuals
-    cat("Scaled residuals", fill = T)
+    cat("Scaled Residuals", fill = T)
     cat("----------------", fill = T)
     resiudal_fivenum <- fivenum(x$residuals)
     residual_fivenum_mat <- matrix(resiudal_fivenum, nrow = 1)
@@ -631,11 +724,11 @@ logLik.hgwrm <- function(object, ...) {
     n <- object$frame.parsed$y
     p <- length(object$beta)
     q <- ncol(object$mu)
-    enp_gwr <- 2 * object$trS[1] - object$trS[2]
+    enp_gwr <- object$trS[1]
     enp_hlm <- p + q * (q + 1) / 2 + 1
     enp <- enp_gwr + enp_hlm
     val <- object$logLik
-    attr(val, "df") <- enp
+    attr(val, "df") <- object$enp
     attr(val, "nall") <- n
     attr(val, "nobs") <- n
     class(val) <- "logLik"