+++ src/RcppExports.cpp
@@ -12,8 +12,8 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // spatial_hetero_perm
-List spatial_hetero_perm(const arma::mat& x, const arma::mat& uv, int poly, int resample, double bw, int kernel);
-RcppExport SEXP _hgwrr_spatial_hetero_perm(SEXP xSEXP, SEXP uvSEXP, SEXP polySEXP, SEXP resampleSEXP, SEXP bwSEXP, SEXP kernelSEXP) {
+List spatial_hetero_perm(const arma::mat& x, const arma::mat& uv, int poly, int resample, double bw, int kernel, int verbose);
+RcppExport SEXP _hgwrr_spatial_hetero_perm(SEXP xSEXP, SEXP uvSEXP, SEXP polySEXP, SEXP resampleSEXP, SEXP bwSEXP, SEXP kernelSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -23,13 +23,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type resample(resampleSEXP);
     Rcpp::traits::input_parameter< double >::type bw(bwSEXP);
     Rcpp::traits::input_parameter< int >::type kernel(kernelSEXP);
-    rcpp_result_gen = Rcpp::wrap(spatial_hetero_perm(x, uv, poly, resample, bw, kernel));
+    Rcpp::traits::input_parameter< int >::type verbose(verboseSEXP);
+    rcpp_result_gen = Rcpp::wrap(spatial_hetero_perm(x, uv, poly, resample, bw, kernel, verbose));
     return rcpp_result_gen;
 END_RCPP
 }
 // hgwr_bfml
-List hgwr_bfml(const arma::mat& g, const arma::mat& x, const arma::mat& z, const arma::vec& y, const arma::mat& u, const arma::vec& group, double bw, size_t kernel, double alpha, double eps_iter, double eps_gradient, size_t max_iters, size_t max_retries, size_t ml_type, size_t verbose);
-RcppExport SEXP _hgwrr_hgwr_bfml(SEXP gSEXP, SEXP xSEXP, SEXP zSEXP, SEXP ySEXP, SEXP uSEXP, SEXP groupSEXP, SEXP bwSEXP, SEXP kernelSEXP, SEXP alphaSEXP, SEXP eps_iterSEXP, SEXP eps_gradientSEXP, SEXP max_itersSEXP, SEXP max_retriesSEXP, SEXP ml_typeSEXP, SEXP verboseSEXP) {
+List hgwr_bfml(const arma::mat& g, const arma::mat& x, const arma::mat& z, const arma::vec& y, const arma::mat& u, const arma::vec& group, double bw, int bw_optim, size_t kernel, double alpha, double eps_iter, double eps_gradient, size_t max_iters, size_t max_retries, size_t ml_type, size_t verbose);
+RcppExport SEXP _hgwrr_hgwr_bfml(SEXP gSEXP, SEXP xSEXP, SEXP zSEXP, SEXP ySEXP, SEXP uSEXP, SEXP groupSEXP, SEXP bwSEXP, SEXP bw_optimSEXP, SEXP kernelSEXP, SEXP alphaSEXP, SEXP eps_iterSEXP, SEXP eps_gradientSEXP, SEXP max_itersSEXP, SEXP max_retriesSEXP, SEXP ml_typeSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -40,6 +41,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::mat& >::type u(uSEXP);
     Rcpp::traits::input_parameter< const arma::vec& >::type group(groupSEXP);
     Rcpp::traits::input_parameter< double >::type bw(bwSEXP);
+    Rcpp::traits::input_parameter< int >::type bw_optim(bw_optimSEXP);
     Rcpp::traits::input_parameter< size_t >::type kernel(kernelSEXP);
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
     Rcpp::traits::input_parameter< double >::type eps_iter(eps_iterSEXP);
@@ -48,14 +50,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< size_t >::type max_retries(max_retriesSEXP);
     Rcpp::traits::input_parameter< size_t >::type ml_type(ml_typeSEXP);
     Rcpp::traits::input_parameter< size_t >::type verbose(verboseSEXP);
-    rcpp_result_gen = Rcpp::wrap(hgwr_bfml(g, x, z, y, u, group, bw, kernel, alpha, eps_iter, eps_gradient, max_iters, max_retries, ml_type, verbose));
+    rcpp_result_gen = Rcpp::wrap(hgwr_bfml(g, x, z, y, u, group, bw, bw_optim, kernel, alpha, eps_iter, eps_gradient, max_iters, max_retries, ml_type, verbose));
     return rcpp_result_gen;
 END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_hgwrr_spatial_hetero_perm", (DL_FUNC) &_hgwrr_spatial_hetero_perm, 6},
-    {"_hgwrr_hgwr_bfml", (DL_FUNC) &_hgwrr_hgwr_bfml, 15},
+    {"_hgwrr_spatial_hetero_perm", (DL_FUNC) &_hgwrr_spatial_hetero_perm, 7},
+    {"_hgwrr_hgwr_bfml", (DL_FUNC) &_hgwrr_hgwr_bfml, 16},
     {NULL, NULL, 0}
 };