+++ src/libhgwr/include/hlmgwr.h
@@ -31,6 +31,22 @@ public:  // Type defs
     };
 
     typedef arma::vec (*GWRKernelFunctionSquared)(arma::vec, double);
+
+    enum class BwOptimCriterionType
+    {
+        CV,
+        AIC
+    };
+
+    typedef double (*BwOptimCriterion)(double, void*);
+
+    static double actual_bw(arma::vec d, double bw)
+    {
+        arma::vec ds = sort(d);
+        double b0 = floor(bw), bx = bw - b0;
+        double d0 = ds(int(b0) - 1), d1 = ds(int(b0));
+        return d0 + (d1 - d0) * bx;
+    }
     
     static arma::vec gwr_kernel_gaussian2(arma::vec dist2, double bw2)
     {
@@ -84,7 +100,24 @@ public:  // Type defs
         double bw;
     };
 
-    using BwSelectionArgs = std::pair<std::reference_wrapper<arma::mat>, std::reference_wrapper<arma::vec>>;
+    // using BwSelectionArgs = std::pair<std::reference_wrapper<arma::mat>, std::reference_wrapper<arma::vec>>;
+    struct BwSelectionArgs
+    {
+        std::reference_wrapper<arma::mat> Vig;
+        std::reference_wrapper<arma::vec> Viy;
+        std::reference_wrapper<arma::mat> G;
+        std::reference_wrapper<arma::mat> u;
+        arma::mat* Ygf;
+        arma::mat* Zf;
+        std::reference_wrapper<arma::mat> mu;
+        std::reference_wrapper<arma::rowvec> rVsigma;
+        std::reference_wrapper<arma::uvec> group;
+        GWRKernelFunctionSquared kernel;
+    };
+
+    static double bw_criterion_cv(double bw, void* params);
+
+    static double bw_criterion_aic(double bw, void* params);
 
 public:
     explicit HGWR(const arma::mat& G, const arma::mat& X, const arma::mat& Z, const arma::vec& y, const arma::mat& u, const arma::uvec& group)
@@ -119,14 +152,14 @@ public:
     explicit HGWR(const arma::mat& G, const arma::mat& X, const arma::mat& Z, const arma::vec& y, const arma::mat& u, const arma::uvec& group, KernelType kernel)
         : HGWR(G, X, Z, y, u, group)
     {
-        this->kernel = kernel;
+        set_kernel(kernel);
         bw_optim = true;
     }
 
     explicit HGWR(const arma::mat& G, const arma::mat& X, const arma::mat& Z, const arma::vec& y, const arma::mat& u, const arma::uvec& group, KernelType kernel, const Options& options)
         : HGWR(G, X, Z, y, u, group, options)
     {
-        this->kernel = kernel;
+        set_kernel(kernel);
         bw_optim = true;
     }
 
@@ -140,7 +173,7 @@ public:
     explicit HGWR(const arma::mat& G, const arma::mat& X, const arma::mat& Z, const arma::vec& y, const arma::mat& u, const arma::uvec& group, KernelType kernel, double bw, const Options& options)
         : HGWR(G, X, Z, y, u, group, options)
     {
-        this->kernel = kernel;
+        set_kernel(kernel);
         this->bw = bw;
         bw_optim = false;
     }
@@ -196,6 +229,21 @@ public:
         }
     }
 
+    BwOptimCriterionType get_bw_criterion_type() { return bw_criterion_type; }
+    void set_bw_criterion_type(BwOptimCriterionType value)
+    {
+        bw_criterion_type = value;
+        switch (bw_criterion_type)
+        {
+        case BwOptimCriterionType::AIC:
+            bw_criterion = &bw_criterion_aic;
+            break;
+        default:
+            bw_criterion = &bw_criterion_cv;
+            break;
+        }
+    }
+
     double get_alpha() { return alpha; }
     void set_alpha(double value) { alpha = value; }
     
@@ -218,6 +266,8 @@ public:
     void set_ml_type(size_t value) { ml_type = value; }
     
     arma::mat get_gamma() { return gamma; }
+
+    arma::mat get_gamma_se() { return gamma_se; }
     
     arma::mat get_beta() { return beta; }
     
@@ -235,9 +285,24 @@ public:
 
     void set_printer(PrintFunction printer) { pcout = printer; }
 
+    double gamma_enp()
+    {
+        double n = 2.0 * trS[0] - trS[1];
+        return n > 0 ? n : trS[0];
+    }
+
+    double enp()
+    {
+        return gamma_enp() + double(nvx + (nvz * (nvz + 1) / 2) + 1);
+    }
+
+    double edf()
+    {
+        return double(ndata) - enp();
+    }
+
 public:
-    double criterion_bw(double bw, const BwSelectionArgs& args);
-    double golden_selection(const double lower, const double upper, const bool adaptive, const BwSelectionArgs& args);
+    int bw_optimisation(double lower, double upper, const BwSelectionArgs* args);
     void fit_gwr();
     arma::vec fit_gls();
     double fit_D(ML_Params* params);
@@ -258,10 +323,11 @@ private:
     double bw = 0.0;
     bool bw_optim = false;
     KernelType kernel = KernelType::GAUSSIAN;
-    GWRKernelFunctionSquared gwr_kernel = &gwr_kernel_gaussian2;
+    BwOptimCriterionType bw_criterion_type = BwOptimCriterionType::CV;
 
     /* model parameters */
     arma::mat gamma;
+    arma::mat gamma_se;
     arma::vec beta;
     arma::mat mu;
     arma::mat D;
@@ -271,6 +337,7 @@ private:
     double alpha = 0.01;
     double eps_iter = 1e-6;
     double eps_gradient = 1e-6;
+    size_t max_bw_iters = (size_t)100;
     size_t max_iters = (size_t)1e6;
     size_t max_retries = (size_t)10;
     size_t verbose = (size_t)0;
@@ -279,6 +346,8 @@ private:
 
 
     /* others */
+    BwOptimCriterion bw_criterion = &bw_criterion_cv;
+    GWRKernelFunctionSquared gwr_kernel = &gwr_kernel_gaussian2;
     std::unique_ptr<arma::mat[]> Zf;
     std::unique_ptr<arma::mat[]> Xf;
     std::unique_ptr<arma::vec[]> Yf;
@@ -294,8 +363,6 @@ private:
     double loglik = 0;
     arma::vec trS;
     arma::vec var_beta;
-    double enp = 0;
-    double edf = 0;
 };
     
 }